+++ convertirEntero.h
@@ -1,8 +1,8 @@
 #ifndef CONVERTIRENTERO_H_INCLUDED
 #define CONVERTIRENTERO_H_INCLUDED
 
-int * decimalOBaseT10Base(char * n, int * Obase, int * vervose);
+long int * decimalOBaseT10Base(char * n, int * Obase, int * vervose);
 
-char * decimal10BaseTDBase(int * n, int * DBase, int * vervose);
+char * decimal10BaseTDBase(long int * n, int * DBase, int * vervose);
 
 #endif // CONVERTIRENTERO_H_INCLUDED
